pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbt-ai-readiness"
version = "0.1.0"
description = "AI Readiness Autopilot for dbt Cloud — auto-enables MCP, generates docs, recommends tests, and scores projects for AI-powered workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.scripts]
dbt-ai-readiness = "dbt_ai_readiness.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_ai_readiness"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
